refactor: Migrate `GroupsSelectClass` to functional component
