+++ src/components/GroupsSelect/GroupsSelect.jsx
@@ -1,25 +1,17 @@
 import cx from 'classnames'
-import flow from 'lodash/flow'
-import get from 'lodash/get'
 import PropTypes from 'prop-types'
-import React from 'react'
+import React, { useState } from 'react'
 
-import { withClient } from 'cozy-client'
-import { translate } from 'cozy-ui/transpiled/react/I18n'
+import { useClient } from 'cozy-client'
+import { useI18n } from 'cozy-ui/transpiled/react/I18n'
 import SelectBox from 'cozy-ui/transpiled/react/SelectBox'
-import Alerter from 'cozy-ui/transpiled/react/deprecated/Alerter'
 import Overlay from 'cozy-ui/transpiled/react/deprecated/Overlay'
 
 import CustomMenu from './SelectBox/Menu'
 import CustomOption from './SelectBox/Option'
 import CustomSelectContainer from './SelectBox/SelectContainer'
-import { createGroup, updateGroup } from '../../connections/allGroups'
-import {
-  translatedDefaultSelectedGroup,
-  isExistingGroup
-} from '../../helpers/groups'
-import container from '../ContactCard/ContactGroupsContainer'
-import SelectedGroupContext from '../Contexts/SelectedGroup'
+import useGroupsSelect from './useGroupSelect'
+import contactGroupConnect from '../ContactCard/ContactGroupsContainer'
 
 const captureEscapeEvent = e => {
   if (e.key === 'Escape') {
@@ -27,175 +19,102 @@ const captureEscapeEvent = e => {
     e.target.blur()
   }
 }
-export class GroupsSelectClass extends React.Component {
-  static contextType = SelectedGroupContext
-  state = {
+
+export const GroupsSelect = ({
+  allGroups,
+  onGroupCreated,
+  cleanTrashedGroups,
+  onChange,
+  closeMenuOnSelect,
+  value,
+  styles,
+  isMulti,
+  noOptionsMessage,
+  withCheckbox,
+  components,
+  className,
+  menuPosition
+}) => {
+  const { t } = useI18n()
+  const client = useClient()
+  const [{ menuIsOpen, editedGroupId }, setState] = useState({
     menuIsOpen: false,
     editedGroupId: ''
+  })
+  const { createGroup, deleteGroup, renameGroup } = useGroupsSelect({
+    allGroups,
+    onGroupCreated,
+    client,
+    t,
+    cleanTrashedGroups
+  })
+
+  const toggleMenu = () => {
+    setState(prev => ({ ...prev, menuIsOpen: !prev.menuIsOpen }))
   }
 
-  toggleMenu = () => {
-    this.setState(state => ({ menuIsOpen: !state.menuIsOpen }))
+  const closeMenu = () => {
+    setState(prev => ({ ...prev, menuIsOpen: false }))
   }
 
-  closeMenu = () => {
-    this.setState({ menuIsOpen: false })
+  const setEditedGroupId = id => {
+    setState(prev => ({ ...prev, editedGroupId: id }))
   }
 
-  setEditedGroupId = id => this.setState({ editedGroupId: id })
-
-  createGroup = async group => {
-    const { allGroups, onGroupCreated, client } = this.props
-
-    if (isExistingGroup(allGroups, group)) {
-      return Alerter.error('groups.already_exists', { name: group.name })
-    }
-
-    try {
-      const { data: createdGroup } = await createGroup(client, group)
-      if (onGroupCreated) {
-        onGroupCreated(createdGroup)
-      }
-      return Alerter.success('groups.created.success')
-    } catch {
-      return Alerter.error('groups.created.error')
-    }
-  }
-
-  deleteGroup = async group => {
-    const { t, cleanTrashedGroups, client } = this.props
-    const { selectedGroup, setSelectedGroup } = this.context
-    const isDeletedGroupSelected =
-      get(group, '_id') === get(selectedGroup, '_id')
-    const { data: flaggedGroup } = await updateGroup(client, {
-      ...group,
-      trashed: true
-    })
-    const alertDuration = 3 * 1000
-
-    const alertTimeout = setTimeout(() => {
-      cleanTrashedGroups()
-    }, alertDuration)
-
-    Alerter.info('groups.removed', {
-      name: flaggedGroup.name,
-      buttonText: t('cancel'),
-      buttonAction: () => {
-        clearTimeout(alertTimeout)
-        this.cancelGroupDelete(flaggedGroup)
-      },
-      duration: alertDuration
-    })
-
-    if (isDeletedGroupSelected) {
-      setSelectedGroup(translatedDefaultSelectedGroup(t))
-    }
-  }
-
-  cancelGroupDelete = async group => {
-    const { client } = this.props
-    delete group.trashed
-    await updateGroup(client, group)
-    Alerter.info('groups.remove_canceled', { name: group.name })
-  }
-
-  renameGroup = async (groupId, newName) => {
-    const { selectedGroup, setSelectedGroup } = this.context
-    const { allGroups, client } = this.props
-    const group = allGroups.find(group => group.id === groupId)
-    const allOtherGroups = allGroups.filter(group => group.id !== groupId)
-    const isRenamedGroupSelected =
-      get(group, '_id') === get(selectedGroup, '_id')
-
-    if (isExistingGroup(allOtherGroups, { name: newName })) {
-      return Alerter.error('groups.already_exists', { name: newName })
-    }
-
-    try {
-      const { data } = await updateGroup(client, { ...group, name: newName })
-      if (isRenamedGroupSelected) {
-        setSelectedGroup(data)
-      }
-      return Alerter.success('groups.renamed.success')
-    } catch {
-      return Alerter.error('groups.renamed.error')
-    }
-  }
-
-  handleChange = props => {
-    const { onChange, closeMenuOnSelect } = this.props
-
+  const handleChange = props => {
     if (closeMenuOnSelect) {
-      this.closeMenu()
+      closeMenu()
     }
 
     onChange(props)
   }
 
-  render() {
-    const {
-      value,
-      allGroups,
-      styles,
-      isMulti,
-      noOptionsMessage,
-      withCheckbox,
-      components,
-      className,
-      closeMenuOnSelect,
-      menuPosition
-    } = this.props
-    const { menuIsOpen, editedGroupId } = this.state
-    const { toggleMenu, setEditedGroupId, handleChange } = this
-
-    const defaultComponents = {
-      Menu: CustomMenu,
-      Option: CustomOption,
-      SelectContainer: CustomSelectContainer
-    }
-
-    return (
-      <div className={cx('u-flex-auto u-w-100', className)}>
-        {menuIsOpen && (
-          <Overlay className="overlay-creation-group" onClick={toggleMenu} />
-        )}
-        <SelectBox
-          classNamePrefix="react-select"
-          isMulti={isMulti}
-          withCheckbox={withCheckbox}
-          menuIsOpen={menuIsOpen}
-          blurInputOnSelect={true}
-          hideSelectedOptions={false}
-          isSearchable={false}
-          isClearable={false}
-          closeMenuOnSelect={closeMenuOnSelect}
-          tabSelectsValue={false}
-          onKeyDown={captureEscapeEvent}
-          noOptionsMessage={noOptionsMessage}
-          options={allGroups}
-          value={value}
-          onChange={handleChange}
-          getOptionLabel={group => group.name}
-          getOptionValue={group => group._id}
-          components={{ ...defaultComponents, ...components }}
-          createGroup={this.createGroup}
-          deleteGroup={this.deleteGroup}
-          renameGroup={this.renameGroup}
-          styles={styles}
-          onControlClicked={toggleMenu}
-          setEditedGroupId={setEditedGroupId}
-          editedGroupId={editedGroupId}
-          menuPosition={menuPosition}
-        />
-      </div>
-    )
+  const defaultComponents = {
+    Menu: CustomMenu,
+    Option: CustomOption,
+    SelectContainer: CustomSelectContainer
   }
+
+  return (
+    <div className={cx('u-flex-auto u-w-100', className)}>
+      {menuIsOpen && (
+        <Overlay className="overlay-creation-group" onClick={toggleMenu} />
+      )}
+      <SelectBox
+        classNamePrefix="react-select"
+        isMulti={isMulti}
+        withCheckbox={withCheckbox}
+        menuIsOpen={menuIsOpen}
+        blurInputOnSelect={true}
+        hideSelectedOptions={false}
+        isSearchable={false}
+        isClearable={false}
+        closeMenuOnSelect={closeMenuOnSelect}
+        tabSelectsValue={false}
+        onKeyDown={captureEscapeEvent}
+        noOptionsMessage={noOptionsMessage}
+        options={allGroups}
+        value={value}
+        onChange={handleChange}
+        getOptionLabel={group => group.name}
+        getOptionValue={group => group._id}
+        components={{ ...defaultComponents, ...components }}
+        createGroup={createGroup}
+        deleteGroup={deleteGroup}
+        renameGroup={renameGroup}
+        styles={styles}
+        onControlClicked={toggleMenu}
+        setEditedGroupId={setEditedGroupId}
+        editedGroupId={editedGroupId}
+        menuPosition={menuPosition}
+      />
+    </div>
+  )
 }
 
-GroupsSelectClass.propTypes = {
+GroupsSelect.propTypes = {
   cleanTrashedGroups: PropTypes.func.isRequired,
   allGroups: PropTypes.array.isRequired,
-  t: PropTypes.func.isRequired,
   styles: PropTypes.object,
   onChange: PropTypes.func.isRequired,
   // for multiple selections, value can be an array
@@ -215,16 +134,14 @@ GroupsSelectClass.propTypes = {
   menuPosition: PropTypes.oneOf(['fixed', 'absolute'])
 }
 
-GroupsSelectClass.defaultProps = {
+GroupsSelect.defaultProps = {
   isMulti: false,
   components: {},
   closeMenuOnSelect: false
 }
 
-const GroupsSelect = flow(withClient, translate(), container)(GroupsSelectClass)
-
 GroupsSelect.propTypes = {
   allGroups: PropTypes.array.isRequired
 }
 
-export default GroupsSelect
+export default contactGroupConnect(GroupsSelect)
